.claude/mcp-servers/agent-chat/requirements.txt
mcp[cli]>=1.0.0
nats-py>=2.6.0
